# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fluxo Inteligente - Sistema de Controle do Refeitório Escolar
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Class {
  id: number;
  name: string;
  year: string;
  /** @nullable */
  studentCount?: number | null;
}

export interface ClassInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  year: string;
}

export interface Student {
  id: number;
  name: string;
  registration: string;
  classId: number;
  className: string;
  /** @nullable */
  qrCode?: string | null;
}

export interface StudentInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  registration: string;
  classId: number;
}

export interface StudentUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  registration?: string;
  classId?: number;
}

export interface MealSlot {
  id: number;
  classId: number;
  className: string;
  startTime: string;
  endTime: string;
  weekDays: string;
}

export interface MealSlotInput {
  classId: number;
  startTime: string;
  endTime: string;
  weekDays: string;
}

export interface MealSlotUpdate {
  classId?: number;
  startTime?: string;
  endTime?: string;
  weekDays?: string;
}

export type CheckinStatus = typeof CheckinStatus[keyof typeof CheckinStatus];


export const CheckinStatus = {
  allowed: 'allowed',
  denied: 'denied',
  no_slot: 'no_slot',
} as const;

export interface Checkin {
  id: number;
  studentId: number;
  studentName: string;
  className: string;
  /** @nullable */
  mealSlotId?: number | null;
  checkedInAt: string;
  status: CheckinStatus;
}

export interface ScanInput {
  studentId: number;
}

export type ScanResultStatus = typeof ScanResultStatus[keyof typeof ScanResultStatus];


export const ScanResultStatus = {
  allowed: 'allowed',
  denied: 'denied',
  no_slot: 'no_slot',
} as const;

export interface ScanResult {
  status: ScanResultStatus;
  message: string;
  studentName: string;
  className: string;
  /** @nullable */
  checkinId?: number | null;
}

export interface DashboardStats {
  totalStudents: number;
  totalCheckinsTodayCount: number;
  totalClasses: number;
  checkinsThisWeek: number;
  /** @nullable */
  mostActiveClass: string | null;
}

export interface SlotWithCount {
  id: number;
  classId: number;
  className: string;
  startTime: string;
  endTime: string;
  weekDays: string;
  checkinCount: number;
}

export type ListStudentsParams = {
classId?: number;
search?: string;
};

export type ListCheckinsParams = {
date?: string;
studentId?: number;
};