# lib/api-zod/src/generated/types/checkin.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fluxo Inteligente - Sistema de Controle do Refeitório Escolar
 * OpenAPI spec version: 0.1.0
 */
import type { CheckinStatus } from './checkinStatus';

export interface Checkin {
  id: number;
  studentId: number;
  studentName: string;
  className: string;
  /** @nullable */
  mealSlotId?: number | null;
  checkedInAt: string;
  status: CheckinStatus;
}