# lib/api-zod/src/generated/types/mealSlot.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fluxo Inteligente - Sistema de Controle do Refeitório Escolar
 * OpenAPI spec version: 0.1.0
 */

export interface MealSlot {
  id: number;
  classId: number;
  className: string;
  startTime: string;
  endTime: string;
  weekDays: string;
}