# lib/api-zod/src/generated/types/mealSlotUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fluxo Inteligente - Sistema de Controle do Refeitório Escolar
 * OpenAPI spec version: 0.1.0
 */

export interface MealSlotUpdate {
  classId?: number;
  startTime?: string;
  endTime?: string;
  weekDays?: string;
}