# lib/api-zod/src/generated/types/scanResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fluxo Inteligente - Sistema de Controle do Refeitório Escolar
 * OpenAPI spec version: 0.1.0
 */
import type { ScanResultStatus } from './scanResultStatus';

export interface ScanResult {
  status: ScanResultStatus;
  message: string;
  studentName: string;
  className: string;
  /** @nullable */
  checkinId?: number | null;
}