# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fluxo Inteligente - Sistema de Controle do Refeitório Escolar
 * OpenAPI spec version: 0.1.0
 */

export interface Student {
  id: number;
  name: string;
  registration: string;
  classId: number;
  className: string;
  /** @nullable */
  qrCode?: string | null;
}