# lib/api-zod/src/generated/types/studentUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fluxo Inteligente - Sistema de Controle do Refeitório Escolar
 * OpenAPI spec version: 0.1.0
 */

export interface StudentUpdate {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  registration?: string;
  classId?: number;
}